pyproject.toml
[project]
name = "ninja-practice"
version = "0.1.0"
description = ""
authors = [
    {name = "Carol",email = "awdrg1210@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "django-ninja (>=1.4.3,<2.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.13"
pre-commit = "^4.2.0"

[tool.ruff]
line-length = 120
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]

[tool.ruff.format]
quote-style = "double"
